test(vllm-frontend): lift bridge demux tests to sibling file, cull low-value cases
bridge.rs was 1059 lines — over the 1k limit — with ~485 lines of inline
demux tests. Lift them to src/bridge/tests.rs (flat layout); the production
code is byte-for-byte unchanged and bridge.rs is now 575 lines.

These are not mock-heavy UTs: they feed real TokenEvents through the real
dispatch_burst/reduce_request and read real EngineCoreOutputs. They stay at
this level because the demux's invariants are not observable at the HTTP
layer the sim frontend_e2e drives — vLLM's UsageInfo is built from raw
prompt/completion counts and never reads prefill_stats, so cached/computed
accounting surfaces only here and in Prometheus, and N->1 batching + abort
are internal-wire behavior. The happy path stays covered end to end by
openinfer-sim's frontend_e2e (the CI gate).

Cull 3 cases that don't raise correctness confidence:
- tokens_then_finish_in_separate_bursts: streaming outcome already covered
  by frontend_e2e ([DONE] after tokens)
- mixed_logprob_batch_keeps_token_alignment: logprobs are uniform per
  request, so the mixed case is unreachable; alignment is structural
- local_ipc_namespace_uses_short_path: frontend_e2e connects over an ipc
  endpoint built the same way, covering the realistic case

Keep 7, each pinning an invariant nothing else covers: reject->error,
token+finish coalesce + logprob batching, first-token metadata once +
computed=prompt-cached, cross-burst deferred flush, stop-on-prefill usage
accounting, multi-request bucketing, and abort dropping late tokens.
