+++ openinfer-vllm-frontend/src/bridge.rs
@@ -572,488 +572,4 @@ async fn wait_for_ipc_endpoint(address: &str, shutdown: &CancellationToken) -> R
 }
 
 #[cfg(test)]
-mod tests {
-    use openinfer_engine::engine::{FinishReason, TokenLogprob};
-
-    use super::*;
-
-    /// Test harness that exercises the bridge's demux path directly: register
-    /// requests, emit tagged events onto the shared channel, drain one ready
-    /// burst at a time, and inspect the coalesced outputs — the same flow the
-    /// `run` loop drives, minus the sockets.
-    struct Demux {
-        event_tx: mpsc::UnboundedSender<(RequestTag, TokenEvent)>,
-        event_rx: TokenStreamReceiver,
-        streams: HashMap<RequestTag, RequestStreamState>,
-        output_tx: mpsc::UnboundedSender<EngineCoreOutputs>,
-        output_rx: mpsc::UnboundedReceiver<EngineCoreOutputs>,
-    }
-
-    impl Demux {
-        fn new() -> Self {
-            let (event_tx, event_rx) = mpsc::unbounded_channel();
-            let (output_tx, output_rx) = mpsc::unbounded_channel();
-            Self {
-                event_tx,
-                event_rx,
-                streams: HashMap::new(),
-                output_tx,
-                output_rx,
-            }
-        }
-
-        /// Register a request as `start_request` does and return its cancel flag.
-        fn add(&mut self, id: &str) -> Arc<AtomicBool> {
-            let tag: RequestTag = Arc::from(id);
-            let cancelled = Arc::new(AtomicBool::new(false));
-            self.streams.insert(
-                Arc::clone(&tag),
-                RequestStreamState::new(Arc::clone(&cancelled)),
-            );
-            cancelled
-        }
-
-        fn emit(&self, id: &str, event: TokenEvent) {
-            self.event_tx
-                .send((Arc::from(id), event))
-                .expect("emit token event");
-        }
-
-        /// Process one ready burst. Returns false if nothing was queued.
-        fn drain(&mut self) -> bool {
-            match self.event_rx.try_recv() {
-                Ok(first) => {
-                    dispatch_burst(
-                        first,
-                        &mut self.event_rx,
-                        &mut self.streams,
-                        &self.output_tx,
-                    )
-                    .expect("dispatch burst");
-                    true
-                }
-                Err(_) => false,
-            }
-        }
-
-        fn next_output(&mut self) -> Option<EngineCoreOutputs> {
-            self.output_rx.try_recv().ok()
-        }
-    }
-
-    #[test]
-    fn rejected_request_is_reported_as_error() {
-        let mut d = Demux::new();
-        d.add("req-1");
-        d.emit(
-            "req-1",
-            TokenEvent::Rejected {
-                message: "request is too large for KV cache".to_string(),
-                prompt_tokens: 16,
-                completion_tokens: 0,
-            },
-        );
-        assert!(d.drain());
-
-        let outputs = d.next_output().expect("terminal output");
-        assert!(
-            outputs
-                .finished_requests
-                .as_ref()
-                .is_some_and(|requests| requests.contains("req-1"))
-        );
-        assert_eq!(outputs.outputs.len(), 1);
-        let output = &outputs.outputs[0];
-        assert_eq!(output.request_id, "req-1");
-        assert_eq!(output.finish_reason, Some(EngineCoreFinishReason::Error));
-        assert_eq!(
-            output.stop_reason,
-            Some(StopReason::Text(
-                "request is too large for KV cache".to_string()
-            ))
-        );
-        assert!(d.next_output().is_none());
-        assert!(
-            !d.streams.contains_key("req-1"),
-            "finished stream is removed"
-        );
-    }
-
-    /// Token(s) and the terminal arriving in one burst (EmitAndFinish) coalesce
-    /// into a single output carrying both the tokens and the finish reason —
-    /// the canonical vLLM shape, one wire message instead of two.
-    #[test]
-    fn token_and_finish_in_one_burst_coalesce() {
-        let mut d = Demux::new();
-        d.add("req-1");
-        d.emit(
-            "req-1",
-            TokenEvent::Scheduled {
-                queued_at_unix_s: 1.0,
-                scheduled_at_unix_s: 2.0,
-                prompt_tokens: 16,
-                cached_tokens: 0,
-            },
-        );
-        d.emit(
-            "req-1",
-            TokenEvent::Token {
-                id: 11,
-                logprob: Some(TokenLogprob {
-                    logprob: -0.1,
-                    top_logprobs: vec![(11, -0.1), (12, -0.5)],
-                }),
-            },
-        );
-        d.emit(
-            "req-1",
-            TokenEvent::Token {
-                id: 21,
-                logprob: Some(TokenLogprob {
-                    logprob: -0.2,
-                    top_logprobs: vec![(21, -0.2), (22, -0.6)],
-                }),
-            },
-        );
-        d.emit(
-            "req-1",
-            TokenEvent::Finished {
-                finish_reason: FinishReason::Length,
-                prompt_tokens: 16,
-                completion_tokens: 2,
-            },
-        );
-        assert!(d.drain());
-
-        let outputs = d.next_output().expect("coalesced output");
-        assert_eq!(outputs.outputs.len(), 1);
-        let output = &outputs.outputs[0];
-        assert_eq!(output.request_id, "req-1");
-        assert_eq!(output.new_token_ids, vec![11, 21]);
-        assert_eq!(output.finish_reason, Some(EngineCoreFinishReason::Length));
-        assert!(output.events.is_some());
-        assert!(output.prefill_stats.is_some());
-        assert!(
-            outputs
-                .finished_requests
-                .as_ref()
-                .is_some_and(|requests| requests.contains("req-1"))
-        );
-
-        let direct = match output.new_logprobs.as_ref().expect("batched logprobs") {
-            MaybeWireLogprobs::Direct(direct) => direct,
-            MaybeWireLogprobs::Wire(_) => panic!("expected direct batched logprobs"),
-        };
-        assert_eq!(direct.positions.len(), 2);
-        assert_eq!(direct.positions[0].entries[0].token_id, 11);
-        assert_eq!(direct.positions[1].entries[0].token_id, 21);
-
-        assert!(d.next_output().is_none());
-    }
-
-    /// Tokens that stream over several steps and a finish that lands later
-    /// arrive in separate bursts: a token output first, a terminal output once
-    /// the finish shows up.
-    #[test]
-    fn tokens_then_finish_in_separate_bursts() {
-        let mut d = Demux::new();
-        d.add("req-1");
-        d.emit(
-            "req-1",
-            TokenEvent::Token {
-                id: 11,
-                logprob: None,
-            },
-        );
-        d.emit(
-            "req-1",
-            TokenEvent::Token {
-                id: 21,
-                logprob: None,
-            },
-        );
-        assert!(d.drain());
-        let token_out = d.next_output().expect("token output");
-        assert_eq!(token_out.outputs[0].new_token_ids, vec![11, 21]);
-        assert!(token_out.outputs[0].finish_reason.is_none());
-        assert!(token_out.finished_requests.is_none());
-
-        d.emit(
-            "req-1",
-            TokenEvent::Finished {
-                finish_reason: FinishReason::Length,
-                prompt_tokens: 16,
-                completion_tokens: 2,
-            },
-        );
-        assert!(d.drain());
-        let terminal = d.next_output().expect("terminal output");
-        assert!(terminal.outputs[0].new_token_ids.is_empty());
-        assert_eq!(
-            terminal.outputs[0].finish_reason,
-            Some(EngineCoreFinishReason::Length)
-        );
-        assert!(
-            terminal
-                .finished_requests
-                .as_ref()
-                .is_some_and(|requests| requests.contains("req-1"))
-        );
-        assert!(d.next_output().is_none());
-    }
-
-    #[test]
-    fn first_token_metadata_is_only_sent_with_first_output() {
-        let mut d = Demux::new();
-        d.add("req-2");
-        d.emit(
-            "req-2",
-            TokenEvent::Scheduled {
-                queued_at_unix_s: 1.0,
-                scheduled_at_unix_s: 2.0,
-                prompt_tokens: 8,
-                cached_tokens: 5,
-            },
-        );
-        d.emit(
-            "req-2",
-            TokenEvent::Token {
-                id: 1,
-                logprob: None,
-            },
-        );
-        assert!(d.drain());
-
-        d.emit(
-            "req-2",
-            TokenEvent::PromptTokens {
-                ids: vec![9],
-                logprobs: vec![None],
-            },
-        );
-        d.emit(
-            "req-2",
-            TokenEvent::Token {
-                id: 2,
-                logprob: None,
-            },
-        );
-        assert!(d.drain());
-
-        let first_batch = d.next_output().expect("first batch");
-        let second_batch = d.next_output().expect("second batch");
-        assert_eq!(first_batch.outputs[0].new_token_ids, vec![1]);
-        assert_eq!(second_batch.outputs[0].new_token_ids, vec![2]);
-        assert!(first_batch.outputs[0].events.is_some());
-        let stats = first_batch.outputs[0]
-            .prefill_stats
-            .as_ref()
-            .expect("first batch carries prefill stats");
-        assert_eq!(stats.num_prompt_tokens, 8);
-        assert_eq!(stats.num_cached_tokens, 5);
-        assert_eq!(stats.num_local_cached_tokens, 5);
-        assert_eq!(
-            stats.num_computed_tokens, 3,
-            "computed must be prompt minus cached, not the full prompt"
-        );
-        assert!(second_batch.outputs[0].events.is_none());
-        assert!(second_batch.outputs[0].prefill_stats.is_none());
-        assert!(d.next_output().is_none());
-    }
-
-    /// A lone `Scheduled` (no token yet) emits nothing; its metadata waits in
-    /// the stream state and flushes onto the first real output.
-    #[test]
-    fn lone_scheduled_defers_until_first_token() {
-        let mut d = Demux::new();
-        d.add("req-defer");
-        d.emit(
-            "req-defer",
-            TokenEvent::Scheduled {
-                queued_at_unix_s: 1.0,
-                scheduled_at_unix_s: 2.0,
-                prompt_tokens: 4,
-                cached_tokens: 0,
-            },
-        );
-        assert!(d.drain());
-        assert!(d.next_output().is_none(), "scheduled alone emits nothing");
-        assert!(d.streams.contains_key("req-defer"), "stream is retained");
-
-        d.emit(
-            "req-defer",
-            TokenEvent::Token {
-                id: 7,
-                logprob: None,
-            },
-        );
-        assert!(d.drain());
-        let output = d.next_output().expect("first token output");
-        assert_eq!(output.outputs[0].new_token_ids, vec![7]);
-        assert!(
-            output.outputs[0].events.is_some(),
-            "deferred scheduled events flush onto the first token"
-        );
-    }
-
-    /// A request that stops on its first sampled token never emits `Token`
-    /// — the terminal output must still deliver the scheduled events and
-    /// prefill stats or cached_tokens silently vanishes from usage.
-    #[test]
-    fn stop_on_prefill_terminal_output_carries_prefill_stats() {
-        let mut d = Demux::new();
-        d.add("req-stop");
-        d.emit(
-            "req-stop",
-            TokenEvent::Scheduled {
-                queued_at_unix_s: 1.0,
-                scheduled_at_unix_s: 2.0,
-                prompt_tokens: 16,
-                cached_tokens: 4,
-            },
-        );
-        d.emit(
-            "req-stop",
-            TokenEvent::Finished {
-                finish_reason: FinishReason::Stop,
-                prompt_tokens: 16,
-                completion_tokens: 0,
-            },
-        );
-        assert!(d.drain());
-
-        let terminal = d.next_output().expect("terminal output");
-        let output = &terminal.outputs[0];
-        assert_eq!(output.finish_reason, Some(EngineCoreFinishReason::Stop));
-        assert!(
-            output.events.is_some(),
-            "queued/scheduled events must flush"
-        );
-        let stats = output
-            .prefill_stats
-            .as_ref()
-            .expect("terminal output must flush prefill stats");
-        assert_eq!(stats.num_cached_tokens, 4);
-        assert_eq!(stats.num_computed_tokens, 12);
-        assert!(d.next_output().is_none());
-    }
-
-    #[test]
-    fn mixed_logprob_batch_keeps_token_alignment() {
-        let mut d = Demux::new();
-        d.add("req-3");
-        d.emit(
-            "req-3",
-            TokenEvent::Token {
-                id: 31,
-                logprob: None,
-            },
-        );
-        d.emit(
-            "req-3",
-            TokenEvent::Token {
-                id: 32,
-                logprob: Some(TokenLogprob {
-                    logprob: -0.3,
-                    top_logprobs: vec![(32, -0.3), (33, -0.7)],
-                }),
-            },
-        );
-        assert!(d.drain());
-
-        let batch = d.next_output().expect("batched output");
-        let direct = match batch.outputs[0]
-            .new_logprobs
-            .as_ref()
-            .expect("batched logprobs")
-        {
-            MaybeWireLogprobs::Direct(direct) => direct,
-            MaybeWireLogprobs::Wire(_) => panic!("expected direct batched logprobs"),
-        };
-
-        assert_eq!(batch.outputs[0].new_token_ids, vec![31, 32]);
-        assert_eq!(direct.positions.len(), 2);
-        assert!(direct.positions[0].entries.is_empty());
-        assert_eq!(direct.positions[1].entries[0].token_id, 32);
-        assert!(d.next_output().is_none());
-    }
-
-    /// Many requests' tokens in one burst coalesce into a single
-    /// `EngineCoreOutputs` with one `EngineCoreOutput` each — the N→1 wire
-    /// collapse that motivated the demux.
-    #[test]
-    fn burst_batches_multiple_requests_into_one_message() {
-        let mut d = Demux::new();
-        d.add("req-a");
-        d.add("req-b");
-        d.emit(
-            "req-a",
-            TokenEvent::Token {
-                id: 1,
-                logprob: None,
-            },
-        );
-        d.emit(
-            "req-b",
-            TokenEvent::Token {
-                id: 2,
-                logprob: None,
-            },
-        );
-        assert!(d.drain());
-
-        let outputs = d.next_output().expect("one coalesced message");
-        assert_eq!(outputs.outputs.len(), 2);
-        let a = outputs
-            .outputs
-            .iter()
-            .find(|o| o.request_id == "req-a")
-            .expect("req-a output");
-        let b = outputs
-            .outputs
-            .iter()
-            .find(|o| o.request_id == "req-b")
-            .expect("req-b output");
-        assert_eq!(a.new_token_ids, vec![1]);
-        assert_eq!(b.new_token_ids, vec![2]);
-        assert!(d.next_output().is_none(), "exactly one wire message");
-    }
-
-    /// After an abort removes the stream entry, a token already in flight for
-    /// that request is dropped instead of producing a stray output.
-    #[test]
-    fn aborted_request_drops_late_tokens() {
-        let mut d = Demux::new();
-        let cancelled = d.add("req-abort");
-
-        // Replicate the Abort handler: flip the cancel flag, drop the stream.
-        cancelled.store(true, Ordering::Relaxed);
-        d.streams.remove("req-abort");
-
-        d.emit(
-            "req-abort",
-            TokenEvent::Token {
-                id: 99,
-                logprob: None,
-            },
-        );
-        assert!(d.drain(), "the late event is consumed");
-        assert!(
-            d.next_output().is_none(),
-            "no output is produced for an aborted request"
-        );
-    }
-
-    #[test]
-    fn local_ipc_namespace_uses_short_path() {
-        let namespace = local_ipc_namespace().expect("create namespace");
-        let input = ipc_endpoint(&namespace, "input.sock");
-        let output = ipc_endpoint(&namespace, "output.sock");
-        assert!(input.len() < 100, "input IPC endpoint is too long: {input}");
-        assert!(
-            output.len() < 100,
-            "output IPC endpoint is too long: {output}"
-        );
-        let _ = std::fs::remove_dir_all(namespace);
-    }
-}
+mod tests;
